+++ src/onegov.ballot/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 - Speeds up clearing election results.
   [msom]
 
+- Presets ballot ids.
+  [msom]
+
 3.7.6 (2019-03-18)
 ~~~~~~~~~~~~~~~~~~~